Fix issues reported by gometalinter, move cli into cmd sub-dir, auto-generate docs for client
